fix: Remove duplicate index route definition
